Deprecate ::LordOfTheRings
