--- doc/unreleased/default/lord_of_the_rings.md
@@ -1,7 +1,0 @@
-# Faker::LordOfTheRings
-
-```ruby
-Faker::LordOfTheRings.character #=> "Legolas"
-
-Faker::LordOfTheRings.location #=> "Helm's Deep"
-```
