+++ doc/unreleased/movies/lord_of_the_rings.md
@@ -1,0 +1,7 @@
+# Faker::Movies::LordOfTheRings
+
+```ruby
+Faker::Movies::LordOfTheRings.character #=> "Legolas"
+
+Faker::Movies::LordOfTheRings.location #=> "Helm's Deep"
+```
